Partway through refactor of Foundation
